Improve downsampling mixed cluster test
