+++ x-pack/plugin/downsample/qa/mixed-cluster/src/yamlRestTest/java/org/elasticsearch/xpack/downsample/MixedClusterDownsampleRestIT.java
@@ -10,8 +10,6 @@ package org.elasticsearch.xpack.downsample;
 import com.carrotsearch.randomizedtesting.annotations.ParametersFactory;
 
 import org.elasticsearch.test.cluster.ElasticsearchCluster;
-import org.elasticsearch.test.cluster.local.distribution.DistributionType;
-import org.elasticsearch.test.cluster.util.Version;
 import org.elasticsearch.test.rest.yaml.ClientYamlTestCandidate;
 import org.elasticsearch.test.rest.yaml.ESClientYamlSuiteTestCase;
 import org.junit.ClassRule;
@@ -19,27 +17,7 @@ import org.junit.ClassRule;
 public class MixedClusterDownsampleRestIT extends ESClientYamlSuiteTestCase {
 
     @ClassRule
-    public static ElasticsearchCluster cluster = buildCluster();
-
-    private static ElasticsearchCluster buildCluster() {
-        Version oldVersion = getOldVersion();
-        var cluster = ElasticsearchCluster.local()
-            .distribution(DistributionType.DEFAULT)
-            .withNode(node -> node.version(getOldVersion()))
-            .withNode(node -> node.version(Version.CURRENT))
-            .setting("xpack.security.enabled", "false")
-            .setting("xpack.license.self_generated.type", "trial");
-
-        if (oldVersion.before(Version.fromString("8.18.0"))) {
-            cluster.jvmArg("-da:org.elasticsearch.index.mapper.DocumentMapper");
-            cluster.jvmArg("-da:org.elasticsearch.index.mapper.MapperService");
-        }
-        return cluster.build();
-    }
-
-    static Version getOldVersion() {
-        return Version.fromString(System.getProperty("tests.old_cluster_version"));
-    }
+    public static ElasticsearchCluster cluster = Clusters.mixedVersionsCluster();
 
     @Override
     protected String getTestRestCluster() {
